+++ lib/company.rb
@@ -12,6 +12,7 @@ require 'company/phone'
 require 'company/resource'
 require 'company/resources/account'
 require 'company/resources/line'
+require 'company/resources/visit'
 require 'company/resources/job'
 
 # After every resource, which is what it hands back.